watch2win-processor-service/pyproject.toml
[project]
name = "watch2win-processor-service"
version = "0.1.0"
description = "Watch2Win search processor service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pydantic-settings>=2.8.1",
    "pymongo>=4.11.3",
    "python-jose[cryptography]>=3.4.0",
    "uvicorn>=0.34.0",
]
